Fix step instrumentation on ARM.
  * A few touchpoints here.  First is the compiler backend side,
in compiler/arm/call, where we finish up the one last piece left
un-done in function calling and we add the missing VOP.  Second is
the compiler common side, where we revert stassats' changes to
disable step instrumentation.  Third is the runtime side where we
tie the trap codes emitted by the compiler into the debugger logic,
including arranging for execution to resume after the end of the
inline trap codes when returning from the trap.

  * While we're here, add a comment to arch_skip_instruction() to
explain where it is lacking as compared with other architectures.

  * And enable step.impure.lisp on ARM.