Slightly unbreak named-readtables per my latest changes.
The compatibility API that is exposed for dispatching characters
can coerce fdefns to symbols, but there is still a problem that
non-dispatching character functions are groveled out and passed
to SET-MACRO-CHARACTER which won't work in general.

Related problem in sb-cover - it did not understand that
GET-MACRO-CHARACTER could return a function-designator.