Don't return (AND ... (NOT SIMPLE-ARRAY)) from TYPE-OF.
Unparsing a CTYPE allows negation unless the result is for TYPE-OF.
Based on an examination of uses of TYPE-OF within the compiler
it did not seem important to have another mode in which TYPE-OF
would be allowed to return negations for internal reasons.

INVALID-ARRAY-INDEX-ERROR uses TYPE-OF on an array, and it's just
as well to avoid mentioning the non-simpleness for readability,
but could easily be restored with (type-specifier (ctype-of ...))
if so desired. INSPECT might want that behavior back as well.

Fixes lp#1317308.