Pedantically replace many uses of :EOF, NIL, and *EOF-OBJECT* with +EOF+.
Unlike in C where the eof value is (supposedly) opaque,
it is silly to provide an EOFP predicate because you still wonder
which particular eof-value it refers to.  *EOF-OBJECT* is not
as obvious as (EQ x +EOF+), and in contrast to LOAD, nothing about
token-level input demands an unforgeable object, so just use 0.