Remove detritus from pretty-printer dispatch mechanism.
COMPUTE-TEST-FN's implementation was largely historical baggage due to
the CONS type specifier only existing in the atom form as far as the
language was concerned. Discriminating on CAR was just for the pprinter.

*PRECOMPILED-PPRINT-DISPATCH-FUNS* didn't work as intended.
It assumed that the first argument to SET-PPRINT-DISPATCH might match a
known entry, but not only was the FUNCTION-CALL type never used for a
builtin entry, it is neither a type-specifier nor external symbol.