Make READ-MAYBE-NOTHING the only right way to call a reader macro.
By eliminating the WITH-CHAR-MACRO-RESULT macro, the functionality of
which is entirely absorbed into READ-MAYBE-NOTHING, alternate
implementations of READ-MAYBE-NOTHING can be considered
- such as alluded to in my recent posting to sbcl-devel - without
reimplementing large parts of the reader as, for instance, in sb-cover.

The present change to sb-cover is just minimally compatible, and
ridiculous, since the entire point of this change is to avoid that.

Also eliminate a needless (DECLARE (SPECIAL)).