adjust the walker's handling of variables a little bit
The walker notes information about variable bindings (and symbol
macros) in the env-lock, but didn't look in the lexenv itself if
the env-lock lookup failed.  This had a bunch of consequences,
notably lp#1368305 and similar.

While we're at it, make the walker aware of global symbol macros
as well (lp#375326)