fix Huffman table build non-determinism
when building, there might be symbols in the symbol code with
the same implied probability (case in point, the #\( and #\)
characters).  We must make sure that those are processed in
the same order no matter the host's hash-table implementation,
so insert enough SORT / STABLE-SORTs to make that happen.

The hash-table manipulations around the Huffman database in
src/code/target-char.lisp do not lead to any non-determinism:
the outputs eventually get sorted by guaranteed-unique keys
(the character code, or the character name's encoding relative
to the now-deterministic Huffman table).