Don't remove safety checks for SLOT-VALUE-USING-CLASS.
SLOT-VALUE-USING-CLASS may be called with a slot definition from a
different class, which is actually sometimes useful, but it may be for
a slot that doesn't exist. And the code for optimized SVUC used
(safety 0), avoiding array bounds checks, causing very hard to track
down issues.
Just remove the safety declaration.