Implement %RAW-INSTANCE-ATOMIC-INCF/WORD interpreter stub.
The vops were already working for PPC and x86 32-bit
but on 64-bit, the vop required a constant index.
But out-of-line calls didn't work on any supported platform.

Also resolve discrepancy between interpreted and compiled code:
because the macroexpansion of ATOMIC-INCF masks the 'delta'
to MOST-POSITIVE-WORD, the out-of-line code must have the identical
Lisp type as the VOP has for the primitive-type and sc restriction,
instead of the user-visible constraint of SIGNED-WORD.