Add support for declaring a build-order file to be built but not loaded.
  * Add a suitable flag for preventing a file from being added to
the list of files to be loaded by genesis, while not preventing it
from being built by the cross-compiler in the first place.

  * This is useful for cases where we want the compile-time side
effects but not the runtime effects of a file, for files that we
want to load into the system another way, and possibly for making
something in the way of a bootstrap test suite (running genesis
multiple times in a session, producing a series of cores based on
a series of otherwise-not-loaded fasls).  In theory, at least.