Make disassembler correctly detect multiple code .ENTRY points
If pad bytes were inadvertently decoded as opcodes, they might bump
the PC beyond where the fun-header-widetag is, and so the .entry hook
would fail to trigger, and everything would be wrong until by chance
the decoder could sync to a valid opcode.

And the use of sb!kernel:code-header-ref in 'add-fun-header-hooks'
was insane. SEG-CODE could be NIL, but it "worked" to dereference NIL
as if a code-header and get back NIL by random coincidence.