Simplify the tests in /CANNOT-SHOW and eliminate an "unreachable code" warning.
This change started out with the intent of just getting rid of the warning
about (VALUES) being dead code, because the preceding ERROR kills it.

Then I observed that a KLUDGE by CSR (ca. 2002) could be removed by moving
string-dispatch later in its file, becoming a def!macro, which needs
sane-package, which is defined above in the same file.

Moreover, two of the failure modes in SUPPRESS-/SHOW-P no longer apply,
as *TRACE-OUTPUT* is set immediately in cold-init.
And since (WRITE `(EXPR= ,(EXPR) ,(OTHEREXPR) ...)) works really early,
for the most part /SHOW and /SHOW0 should be considered deprecated.

It would be great if somebody would make cold-stderr-stream work on win32.