Fix RUN-PROGRAM failing if an output file doesn't exist.
If a pathname is passed as :OUTPUT or as :ERROR to RUN-PROGRAM, the
target file is created if it doesn't exist. The problem was that when
:IF-OUTPUT-EXISTS / :IF-ERROR-EXISTS is also specified, RUN-PROGRAM
signaled an error if the file cannot be found. This behavior is caused
by NIL being passed as :IF-DOES-NOT-EXIST argument to OPEN, so let's fix
it by explicitly passing :IF-DOES-NOT-EXIST :CREATE.