Fix and re-activate wiping of unused words in GC cards that have non-movable words in them.
The fix is switching from using malloc in the GC to using an area
statically mmapped, and "freed" outside GC with madvise(DONTNEED).
The latter is disabled for Win32 but it isn't that much memory.

I did reproduce previous hangs and it did not hang with this diff
in the same run.