Exact arguments for invalid arg-count backtraces on ARM, x86oids.
Move argument checking before the registers and the stack are
overwritten, this allows for the backtrace to recover all the
arguments exactly as they were passed, including the extra ones.

This is actually implemented for all platforms (albeit not tested),
but not all are either interrupt safe in that place or can produce a
proper backtrace from that interrupt, hence :precise-arg-count-error
feature is only enabled on ARM and x86oids.