Remove DUMP-FOP* and DEFINED-CLONED-FOPS.
Cloned pairs no longer exist. Instead, the loader's main loop can
provide 1 or 2 non-stack arguments to a fop function, and each
can be independently stored in the fasl as 1, 2, 4, or 8 bytes.
This is based on using low bits of the opcode in a uniform way.

A bunch of fops were renumbered to make space for the new encodings,
which breaks fasl compatibility badly, but that's never guaranteed
across builds.