Move 2 slots from most instances of CTYPE into their TYPE-CLASS.
ENUMERABLE and MIGHT-CONTAIN-OTHER-TYPES-P are almost always constant
for all instances of a type-class, so by making them into
possibly-constant pseudo-methods in the manner of SINGLETON-P,
the compiler conses about 1% less and is no slower for it.

The cagey remark that ENUMERABLE is "Meaningless in translated classes"
is contradicted by the fact that CHARACTER is painstakingly arranged to be
enumerable unlike any other classoid. If "meaningless" equates to "not used"
or "is bogus, do not look  at it", then why do this? I've no idea, but
that aspect is preserved for no discernable reason except compatibility.

Also make TYPE-CLASS-NAME read-only.