src/compiler/stack: Be more clever about back-propagation of DX LVARs.
  * DX LVARs don't represent values, they represent storage for
values.  In some cases the storage can be conditionally allocated,
while the value itself is still valid (such as in (IF X (CONS ...))
which could lead to not finding a USE of the LVAR along one
backwards path, thus leading to it being considered live on entry
to the function.  Not exactly excellent.

  * Use a depth-first-search to do back-propagation of DX LVARs
from their environment entry to their USEs only along control-flow
paths which actually allocate the LVARs, with appropriate
cleverness to prevent getting stuck in loops or searching "too far
back".

  * This is in support of fixing lp#1044465, but doesn't actually
affect the symptom, let alone cause correct code to be generated.