compiler/stack: Order block UVL sets for the conditional-DX case.
  * When we're dealing with the hard case for conditional DX we
end up with blocks with a start-stack that is the union of the end
stacks of their predecessors, breaking the assumption that any one
predecessor has all of the UVLs on its end-stack.

  * To fix this, detect the situation, decline to order the UVLs
until all of the predecessors have had their UVLs ordered, and
then do an in-order union operation to compute the "effective"
predecessor end stack.

  * And since we've now arranged for a violation of an assumption
in DISCARD-UNUSED-VALUES, add an AVER there to keep from getting
into an infinite loop.  We'll work on that next.

  * This is in support of lp#1044465, finally affecting the
symptom but not yet allowing the code to compile or generating
correct code.