make-host-2 now with even less noise, and like OMG how did this work?
It seems that CTYPEP calls %TYPEP to determine the metatype (the type
of its second argument), because it does not know any of the type-classes
{NUMERIC-TYPE, NAMED-TYPE, MEMBER-TYPE, ...} which actually means we're
hitting a type cache to parse 'SB-KERNEL:NUMERIC-TYPE etc every time.
Infinite recursion is avoided because all type-classes are CLASSOIDs,
and CLASSOID-TYPEP is a base case.  I have a (large) fix for this.