Remove quasi-builtin non-atomic types.
Having (MOD 2) etc. considered builtin is unimportant. It was previously
an assurance that parsing returned the same object every time, but other
mechanisms besides globaldb do that: caching and interning. Interning is
stronger as it equates all spellings of one type, e.g. (MOD 4) is EQ to
(UNSIGNED-BYTE 2), (INTEGER 0 3),  (INTEGER (-1) 3) and infinitely more.