compiler/stack: Bug 308914 workaround should never trigger.
  * This "workaround" is for a dead-code case where STACK analysis
is confused by some loop of dead-but-not-deleted blocks remaining
in the component.  In actual practice, it mostly drives the bugs
involved to become subtler, causing errors and infinite loops
slightly later in STACK analysis.

  * The new DCE pass should prevent this case from recuring, so
killing the workaround should do no harm, though making it into an
AVER rather than merely reverting to an infinite loop is an
improvement.