Fix package locking in WITH-FAST-READ-BYTE.
The macro WITH-FAST-READ-BYTE disables the package lock for the symbol
FAST-READ-BYTE during a certain part of its expansion. The intention
seems to have been to let this part end before the macro's spliced body
but it actually extends over that. Fix this.

This change has no visible effect, it just makes the macro adhere better
to good practice for defensive programming.

For use by the cross compiler there is another version of this macro.
That version is not affected and therefore left unchanged.