Improve SXHASH on bit-vectors of length divisible by the word size.
For these lengths, the implementation so far didn't use the bits of
the final word of the vector's contents to calculate the hash code,
particularly making the value of SXHASH on all word-sized bit-vectors
the same. This seems to have been an oversight, so fix it.

Fix a nearby typo in NEWS.