Don't always load the destination frame pointer in the MOVE-WORD-ARG VOP.
The MOVE-WORD-ARG VOP takes as one argument a pointer to the frame to
which it should move the other argument. The condition under which to
load this frame pointer was (NOT (SC-IS Y SAP-REG)), which is always
true there. This manifested in redundant moves when the value to move
was held in registers in both frames involved. For example on x86-64
near the return from a local function returning an untagged word there
could occur

  MOV RAX, [RBP]

with RAX being dead afterwards.

I guess the MOVE-WORD-ARG VOP was copied from the MOVE-SAP-ARG VOP where
the same condition is (correctly) used, and adapting the condition was
forgotten.

This commit changes, in all backends, the condition in the MOVE-WORD-ARG
VOP to (NOT (SC-IS Y SIGNED-REG UNSIGNED-REG)), thus not loading the
destination frame pointer if the argument move is between registers.

I have tested this change on x86-64 and x86; for the other backends I
only relied on careful editing, and running make-host-1.sh to ensure
that the changed VOP can at least be compiled.