one more genesis leak
I think this relates to lambdas created at genesis time.  They get
given names, but none of the usual sources for names is present,
so it falls back to the host's *load-truename*, which is unhelpful
because the only binding of *load-truename* is for the whole build
script itself, src/cold/compile-cold-sbcl.lisp.  This patch doesn't
fix that but does make the value given independent of build
directory