Remove an ugly hack from PARSE-DEFMACRO.
When &environment is accepted, it should only appear at toplevel
in the lambda-list. This removes the feature of allowing it in
a nested list to support the bogus way that DEFSETF called it.
[Tested by pre-existing "test-defsetf-env-1" and "-2" tests]

Also rename ASSIGN-SETF-MACRO to %DEFSETF for consistency with
other functions that install code created by a defining macro.