Eliminate the badly named 'expand-or-get-setf-inverse' function.
When it returned a call to `(SETF ,SYM) this was not using
an inverse - it's a "setf function name".

And when in the SETF macro the :inverse exists, and the place is
not a locally defined function, GET-SETF-EXPANSION is not needed
to decide for us that the inverse can be used - it can be.