Remove COMPILER-DESTRUCTURING-BIND, just use TYPEP.
Nikodemus invented this in 92d1243a but it is not our usual idiom
for "if a form matches a pattern, then treat it specially."
Macro-like transforms that want to pick off a special case should
do precisely that and no more: accept the recognized case, and let
other things fall through; not pile insanity onto PARSE-DEFMACRO to
tell it to invoke COMPILER-ERROR vs. ERROR vs. ARG-COUNT-ERROR ...