Change MACROLET IR1 translator to use MAKE-MACRO-LAMBDA
There is a minor discrepancy in whether repetition of variables
in a destructuring lambda list is permitted. MAKE-MACRO-LAMBDA
allows it but PARSE-DEFMACRO does not, which is slightly wrong
at least according to a quick survey of other implementations,
most of which permit (X &REST X), though in some cases you'll get
a warning about non-use of the earlier X.

When everything has been converted to MAKE-MACRO-LAMBDA
we should add a NEWS entry documenting the relaxed stance.