Make backend-internal-errors into a compile-time constant.
This change removes a workaround for the fact that cold-load could not
observe some of the error symbols, which, absent the workaround,
would not have been exported from SB-KERNEL. [*]

Also, the sanity check in the DEFERR macro worked only by dumb luck
for lambda lists which contained &OPTIONAL.

[*] A symbol is exported only if it is was external in the host _and_
was the argument to COLD-INTERN at least once. This has the desirable
effect of ignoring obsolete junk in package-data-list as well as symbols
inapplicable to the target. Calling INTERN during cold-init won't
create an external symbol.