Rearrange a bunch of stuff about CTYPEs, LAYOUTs, defstructs.
The net effect of this change is to open-code more structure accessors,
optimally compile just about all calls to TYPEP on CONDITION subtypes,
and print ~1000 lines (= 4%) less of optimization-related noise
to stderr during make-host-1 and make-host-2.
- When building the xc, there are neither effiency notes nor
  style-warnings up to (but excluding) compilation of "compiler/node".
- When running the xc, there are similarly no notes/warnings up to but
  excluding "early-fasl" [ignoring those from loading the cross-compiler]
  and no style-warnings about unknown type ERROR, etc.
- Globaldb requires one fewer evil hack regarding type-check of LAYOUT.

To achieve this, type operations which aren't needed in the host
are moved from 'early-type' to 'target-type'; and CLASSOID-CELL,
CLASSOID, LAYOUT, DEFSTRUCT-DESCRIPTION are defined together,
reducing the number of files involved in a dependency cycle,
and eliminating all but one unoptimizable slot type-check (in the
out-of-line accessor for LAYOUT-CLASSOID, only in the cross-compiler)
that is never hit, so irrelevant.