Simplify MAKE-CONDITION, ALLOCATE-CONDITION.
Distinguishing two allegedly different errors in ALLOCATE-CONDITION
was not helpful. If you invoked (ALLOCATE-CONDITION 5), it said
"5 does not designate a condition class" whereas (ALLOCATE-CONDITION T)
said "T is not a condition class". While either message is right,
the former is nicer and more general because it says "designate."

/SHOW0 in WARN was useless because conditions crash cold-init anyway.

Also, ALLOCATE-INSTANCE should not return 2 values.