fix an xc host difference
The :SCS argument in a VOP ends up in a VOP-PARSE structure, which
then gets dumped.  The :SCS are sanitized by REMOVE-DUPLICATES,
which on SBCL (and CCL) returns a fresh list in all cases, but on
CLISP returns its argument if there are no duplicates.

(This was leading to differences in the FOP sequences in
src/compiler/x86/call.lisp, where the sequence for explicit
structure sharing from CLISP-as-xc was different from the sequence
for constant coalescing from SBCL-as-xc)