Fix multiple-value restoring in the presence of alien calls on x86-64.
Saved multiple-values are pushed onto the stack and the old stack
pointer is saved, upon restoring the size is computed oldSP-SP, but
before calling alien functions the stack pointer is aligned, throwing
off the number of saved values. Store and restore the stack pointer
around alien calls on c-stack-is-control-stack, not just on x86.