Restore sb-grovel backwards compatibility.
In the past sb-grovel treated everything as `int', but now it uses long.
Which caused problems with win64, where things that were expected to
be signed became unsigned.
On x86-64: if a value fits into 4 bytes, cast it to `int', effectively using
(sizeof(x) == 4)? (long) (int) (x): (x)
with obvious differences for long-long on win64.
This restores the previous behaviour while allowing for `long'
constants to be grovelled.