Declare COERCE and two helpers as EXPLICIT-CHECK.
COERCE does not benefit from checking anything in its XEP because
the first thing it does is call SPECIFIER-TYPE which will detect problems.
In the case of coercing to a vector, there were 3 layers of checks on
the specifier: first, was it (OR SYMBOL CONS INSTANCE); second, did it
designate a type; and third, did it designate a sequence.
Instead, just let MAKE-SEQUENCE figure things out.