Fix unreliable PARSE-UNKNOWN-TYPE signaling.
This change skips caching of results when any subpart of a type expression
signals a parse-unknown-type condition. Prior to this change, a condition
would be signaled for UNKNOWN in the expression (ARRAY UNKNOWN),
but the list expression itself was still entered into the cache,
which inhibited further conditions regarding (ARRAY UNKNOWN).

Additionally we now avoid using cache lines for classoids used as types.