Do not compile some pieces of DEFSTRUCT-defined things at fasl load-time.
A few things would formerly be compiled at load-time:
- a thunk for ALLOCATE-INSTANCE
- out-of-line slot setters for read-only slots

This was possibly the worst choice of when to compile.
Not only was the fast loader slowed down for the rarest of use-cases,
it might unexpectedly reference stuff that is as yet undefined,
depending on fasl load order (especially if we can parallelize).

A good choice would be do everything at compile-time, except that's not
how it works. A reasonable fallback is to compile at runtime, obeying the
principle of least surprise compared to all other method dispatch logic.