Add :SB-FASTEVAL feature.
A small README file and NEWS entry will be forthcoming.

The short version is: deleting :SB-EVAL and adding :SB-FASTEVAL
to 'customize-target-features' will get you an SBCL that defaults
to using the new interpreter at the REPL. This is subject to change,
but seems to easiest way to ensure that the contribs run their
tests as interpreted code.

In terms of performance, CL-BENCH indicates speedups relative to
sb-eval on benchmarks anywhere up to 100x (!) depending on how much
it uses the kernel versus the interpreter. On the low end, bignum
tests are unaffected - any apparent difference is just noise.
More typically the benchmarks show 10x to 20x improvement though.

Also run-test.sh should mostly pass, if you're so inclined.