Fix policy-related stuff in the new interpreter.
All along the POLICY macro was turning the NIL environment policy
into **BASELINE-POLICY**, effectively causing the interpreter
to ignore toplevel (DECLAIM) forms.

I'm not sure I agree with the reasoning by which the compiler needs
NIL to mean baseline, but it has to do with asking for the policy
on the dummy head and tail blocks of a component, among other things.

This explains why every time I wanted to benchmark the interpreter,
its source code had to be hacked to ignore all type checks.
Yah, I actually did it that way.