Improve the disaster which is FLOAT-NAN-P and FLOAT-TRAPPING-NAN-P
  * IEEE floats typically use values with all of the exponent bits
set as infinities and NaNs.  Typically, the infinities have all of
the mantissa bits clear, and the NaNs have at least one of them
set.

  * It turns out that many implementations, including all of the
implementations that we support, use the highest mantissa bit to
distinguish between "Quiet NaNs" (QNaNs) and "Signalling NaNs"
(SNaNs, or "trapping NaNs" in our terminology).  The IEEE float
spec apparently requires that there be at least one QNaN and at
least one SNaN, but doesn't mandate a particular format.

  * HPPA and MIPS treat the bit being clear as a QNaN and the bit
being set as an SNaN.  ALPHA, PPC, SPARC, and x86 (and presumably
ARM, ARM64, and x86-64) treat the bit being clear as an SNaN and
the bit being set as a QNaN, hence the conditional logic for
FLOAT-NAN-P and FLOAT-TRAPPING-NAN-P.

  * Unfortunately, the prior effect of FLOAT-NAN-P on HPPA and
MIPS was to return true for infinities OR SNaNs, and false for
numbers and QNaNs, and FLOAT-TRAPPING-NAN-P would return true for
all NaNs.

  * This was compounded by the fact (still the case) that on other
platforms (not HPPA or MIPS), FLOAT-TRAPPING-NAN-P returns true
for infinites and SNaNs, and false for numbers and QNaNs.

  * So, improve the situation as follows: For all platforms, have
FLOAT-NAN-P return true for both SNaNs and QNaNs, and false for
all other values.  For HPPA and MIPS, have FLOAT-TRAPPING-NAN-P
return true for SNaNs, and false for all other values.  For all
other platforms, have FLOAT-TRAPPING-NAN-P return true for SNaNs
and infinities, and false for all other values (preserving the
existing behavior).

  * Two things to follow up on here are fixing the semantics for
FLOAT-TRAPPING-NAN-P to return false for infinities on all
architectures, and pushing the parameterization on the SNaN vs.
QNaN interpretation of the "trapping nan bit" to
SYS:SRC;COMPILER;TARGET;PARMS.LISP.