Define and use new feature, :64-BIT-REGISTERS
  * For those machines (ALPHA, and a hypothetical MIPS n32 port)
that have 64-bit wide registers, but use a 32-bit wide heap word,
we need a feature that isn't :64-BIT, but triggers some of the
behavior that was folded into the :64-BIT feature.

  * Specifically, :64-BIT implies N-WORD-BITS = 64, while
:64-BIT-REGISTERS implies N-MACHINE-WORD-BITS = 64, thus is also
present for all :64-BIT backends.

  * And update those conditionals which once tested
N-MACHINE-WORD-BITS but were converted to test :64-BIT to use
:64-BIT-REGISTERS instead (except for one, which should have used
N-WORD-BITS in the first place).