Part 1 of 2-part fix for lp#309102 "Explicit-check defknowns"
This change adds a local declaration for EXPLICIT-CHECK,
which for the moment is redundant with the globaldb attribute.
(As a sanity check, both must be present or not present.)

Globaldb was ill-suited to represent this aspect of a function.
Explicit-check pertains how the function's body is expressed,
and not the "contract" of the function. If the function was stated
to explicitly check argument/result types, but it wasn't written
that way, it's a problem. A local declaration does not prevent
this from happening, but lessens the chance of it.