Fix position of alien single-float and double-float stack parameters on Alpha
  * Due to register pressure concerns, the fifth and sixth integer
argument registers are passed to call_into_c on the stack, and
call_into_c loads them from the stack, fixes up the stack pointer,
does the call, and reverts the stack pointer.

  * What does this have to do with floating-point stack arguments?
Only this: They must use the same offset within the stack, even if
the decision point for what is a register argument vs. a stack
argument is different from the integer case.