Fix NaN comparisons on Alpha
  * For any float comparison involving a NaN, we define the result
to be NIL (or false).  I don't know what my system is doing, but
it's not doing that by default (could be an incorrect trap mode
for all I know).  The simple thing, though, is to take advantage
of the "unordered" comparison predicate, which is for precisely
this, and add a second branch.  Problem solved.