tests: Fix and annotate tests for (not :stack-allocatable-closures)
  * One test needed to be adjusted to not signal a WARNING or
COMPILER-NOTE with ERROR, but instead wrap it in a SIMPLE-ERROR
(which has some precedent elsewhere, as it tends to stop further
tests from being run when it occurs).

  * For the rest, an x86-64 --without-stack-allocatable-closures
--without-sb-thread build was made, and compared to a build merely
--without-sb-thread, and the failing tests compared against the
(much longer) list of tests failing on hppa (which doesn't yet
have the :stack-allocatable-closures feature by default).  Any
test newly failing on x86-64 and already failing on hppa was then
marked as being skipped if the feature was not present.