hppa: Fix float negation
  * The previous implementation of %NEGATE for floating-point
values simply subtracted the value from zero...  Which worked for
most cases, but when the input was a positive zero then the output
was as well, which is wrong.

  * Arranging to multiply the input by an appropriately formatted
negative one produced the correct result for a zero input...  and
an overflow condition on some other input.

  * PA-RISC 2.0 has an FNEG instruction which should do what we
need, but we're at least making a pretense at PA-RISC 1.1
compatibility, so that's not useful.

  * The "right" thing to do is to export the input from the FPU to
the CPU, flip the sign bit, and then re-import it to the FPU, but
that's two round-trips via a stack slot, plus two CPU registers to
hold the value and the sign-bit to use with XOR, which is overkill
when the common case can be handled with a single FPU instruction.

  * And what I did instead was to check for a zero input, multiply
by negative one if the input is zero, and subtract from zero if
the input is not zero.  The least dreadful of two solutions that
I've been able to find.  Can someone please find a better option?