Use DEFMETHOD for SIMD-PACK printer.
Despite the insidious motive - rather than fixing the not-inlineability
during cross-compilation, side-stepping the issue - it's not a bad thing.

If all of the OUTPUT-mumble functions were methods, and if we could
cross-compile a discriminating function for PRINT-OBJECT so that printing
were not totally incapacitated in cold-init, we'd be on to something.