Fix massively overcomplicated disassem-fun-cache.
The "cache" a/k/a an identical-code-folding optimization, was based on
a not-quite-correct abstract representation of the disassembler's
generated printer functions prior to translating them from their DSL
into Lisp, as in (:name :tab reg) -> (PRINT "op") (TAB) (PRINT reg).

As well as being simpler to compare after conversion to Lisp,
we get better results: (APROPOS-LIST "INST-PRINTER-") used to
list 125 symbols on x86-64, and now it returns only 88 symbols.