Fix an AVER caused by ignoring closed over variables.
The recent optimization where zeroes in (vector 0 0) are not written
confuses the register allocator when (vector x) is in a closure and X
has a type of (EQL 0), since the closure is allocated but the value is
never used.
Only consider strict constants, this is similar to
5b4dae02596a5d6a6c5db9fed2de089d1bd5481f.

Reported by Eric Marsden.