Unintern %empty-ht-slot% and put it near MAPHASH
On x86-64, make it a static symbol to avoid RIP-relative mode
and a code header word. It encodes to the same number of bytes.
For others, do some magic to permit uninterning.

Fwiw, the lengthy remark about how it would be nice to unintern
was no more applicable to this one symbol than anything else,
if DO-ALL-SYMBOLS is your weapon of choice. There are plenty of
other ways to ruin the system, mostly prevented by package locks.