Fix terrible and terribly useless semantics of FUN-INFO-OR-LOSE.
DEFKNOWN tried to share FUN-INFOs. FUN-INFO-OR-LOSE (the "reader")
copied them because it didn't know if a mutator was reading.
But sharing was only possible in the first place if a single DEFKNOWN
defined a bunch of functions together, and by and large,
most of them became non-shared. We can coalesce at image dump time
to achieve the intention, only much better.

And a gratuitous unrelated change: fewer * prompts in dump phase.