Remove remnants of non-interleavable raw slot spport.
This change removes a confusing artifact of trailing raw slot support.
Boxed objects which represent their length in a header word (i.e. not code
or vector) are supposed to use the logical length, not physical length.
You can derive an exact physical word count from the logical length,
but not the other way. Structures had a bunch of off-by-1 length-returning
functions depending what you wanted, all because of trailing raw slots.
Not any more. DD-LENGTH, LAYOUT-LENGTH, and %INSTANCE-LENGTH return identical
values, and DD-LAYOUT-LENGTH and DD-INSTANCE-LENGTH are gone.

A few places that compute a physical length must explicitly account for
padding, if desired. These places are {scav,size,trans}_instance(), two of
which already did that - "scav" didn't - and genesis which already did,
and DO-INSTANCE-TAGGED-SLOT, which now does it a little differently.