Remove :sb-just-dump-it-normally magic.
Now that MAKE-LOAD-FORM-SAVING-SLOTS works in the cross-compiler,
and is turned into FOP-STRUCT, we use that, thereby ensuring that
both the cross-compiler and target compiler work, or both don't work.
It's a maintenance win, even if slightly slower. It could be even
more beautiful if M-L-F-S-S produced the same forms for structure-object
and standard-object, which we could then optimize differently.